Create initial draft for Wrapped Unit Token post.
I've been working on this idea in my spare time for the past month-ish and I'd like to start publishing about it and posting it online. In this commit I am publishing my initial writings about the idea (in the form of a project README) so I can refine it for consumtion on my website and the broader internet.

My goal is to refine this initial documentation of the idea and publish it for consumption, so its posting may be treated about a RFC or whitepaper.